Create README.md
It's priority_queue library, max-heap. At any point of time, It'll help us to find the largest/smallest element from this heap in o(1) time, which is awesome.Internally it's creating a complete binary tree where insertion and deletion is proportional to height of the tree, which is nothing but o(logn).
I designed a templated class so that heap can able to store different data types like int, float and double, and we don't need to write code for different data types.
I also created a useMaxHeap.cpp file for testing this library. Go and check it out.